# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WAFFER",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://qasaedona-api.squaretech.tech/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/admin/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "description": "get all permissions",
                "operationId": "4c0aee37d4eecedce47288592cdd8eb8",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/permissions/me": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "description": "Get my permissions",
                "operationId": "43edc8396c8449d05255fb7301e77d9e",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/roles/{role}/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions for a role",
                "operationId": "9ff29d09d1829574956b9ac88a877549",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Set permissions for a role",
                "operationId": "9d9f1735dffe7bc4a1aed0879cec6920",
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "description": "The ID of the role",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "permissions[0]": {
                                        "description": "The names of the permissions",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "permissions[1]": {
                                        "description": "The names of the permissions",
                                        "type": "string",
                                        "default": ""
                                    },
                                    "permissions[2]": {
                                        "description": "The names of the permissions",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. This action is unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}